miner dashboard restructure
