+++ src/pages/MinerDashboard/MinerDashboard.page.tsx
@@ -17,10 +17,10 @@ import { minerDetailsGet } from 'src/rdx/minerDetails/minerDetails.actions';
 import { minerHeaderStatsGet } from 'src/rdx/minerHeaderStats/minerHeaderStats.actions';
 import { minerStatsGet } from 'src/rdx/minerStats/minerStats.actions';
 import { minerStatsChartGet } from 'src/rdx/minerStatsChart/minerStatsCharts.actions';
-import { AccountHeader } from './components/Header/AccountHeader';
-import { HeaderGreetings } from './components/Header/Greetings';
+import { AccountHeader } from './components/Header/components/AcountHeader/AccountHeader';
+import { HeaderGreetings } from './components/Header/components/Greetings/Greetings';
 import { HeaderStats } from './components/Header/Stats';
-import { MinerDetails } from './components/Header/MinerDetails';
+import { MinerDetails } from './components/Header/components/MinerDetails/MinerDetails';
 import { MinerStatsPage } from './components/Stats/MinerStats.page';
 import { FaChartBar, FaCube, FaWallet } from 'react-icons/fa';
 import { Spacer } from 'src/components/layout/Spacer';
