+++ src/pages/MinerDashboard/components/Blocks/MinerBlocks.page.tsx
@@ -1,6 +1,6 @@
 import { Helmet } from 'react-helmet-async';
 import { useRouteMatch } from 'react-router';
-import { BlocksSection } from './MinerBlocks.section';
+import { BlocksSection } from './components/MinerBlocksSection/MinerBlocks.section';
 export const MinerBlocksPage = () => {
   const {
     params: { address },
