+++ src/pages/MinerDashboard/components/Blocks/components/MinerBlocksSection/types.tsx
@@ -1,0 +1,19 @@
+export type ApiBlock = {
+  confirmed: boolean;
+  difficulty: number;
+  hash: string;
+  luck: number;
+  miner: string;
+  number: number;
+  region: string;
+  reward: number;
+  roundTime: number;
+  timestamp: number;
+  type: 'block' | 'uncle' | 'orphan';
+};
+
+export type ApiBlocks = {
+  totalItems: number;
+  totalPages: number;
+  data: ApiBlock[];
+};
