+++ src/pages/MinerDashboard/components/Header/components/MinerDetails/components.tsx
@@ -1,0 +1,33 @@
+import styled from 'styled-components';
+
+export const NoFeeLimit = styled.div`
+  color: var(--text-secondary);
+`;
+
+export const Item = styled.div`
+  display: flex;
+  font-weight: 600;
+  margin-right: 2rem;
+  white-space: nowrap;
+`;
+
+export const Content = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  @media screen and (max-width: 900px) {
+    margin-top: -0.5rem;
+    ${Item} {
+      margin-top: 0.5rem;
+      flex-grow: 1;
+      flex-shrink: 0;
+      margin-right: 1rem;
+      display: block;
+    }
+  }
+  @media screen and (max-width: 500px) {
+    margin-top: -0.5rem;
+    ${Item} {
+      width: calc(50% - 2rem);
+    }
+  }
+`;
