+++ src/pages/MinerDashboard/components/Payments/components/MinerPaymentsList/MinerPayments.list.tsx
@@ -21,52 +21,8 @@ import {
   useLocalizedCurrencyFormatter,
   useLocalizedNumberFormatter,
 } from 'src/utils/si.utils';
-import styled from 'styled-components';
 import { BiTransferAlt } from 'react-icons/bi';
-
-const HeaderSplit = styled.div`
-  display: flex;
-  justify-content: space-between;
-`;
-
-const StatusContainer = styled.span<{
-  confirmed: ApiMinerPayment['confirmed'];
-}>`
-  display: inline-block;
-  text-transform: capitalize;
-  white-space: nowrap;
-  & + * {
-    margin-left: 0.5rem;
-  }
-  ${(p) =>
-    p.confirmed === true &&
-    `
-      color: var(--success);
-  `}
-  ${(p) =>
-    !p.confirmed &&
-    `
-      color: var(--text-secondary);
-  `}
-  + * svg {
-    fill: var(--text-tertiary);
-  }
-`;
-
-const ButtonDateSwitch = styled(Ws)`
-  padding: 0 0.35rem;
-  outline: none;
-  border: none;
-  color: var(--text-secondary);
-  svg {
-    opacity: 0.5;
-    margin-left: 0.3rem;
-  }
-  &:hover svg {
-    color: var(--primary);
-    opacity: 1;
-  }
-`;
+import { HeaderSplit, StatusContainer, ButtonDateSwitch } from './components';
 
 export const MinerPaymentsList: React.FC<{
   address: string;
