+++ src/pages/MinerDashboard/components/Payments/components/PaymentsChart/Payments.chart.tsx
