+++ src/pages/MinerDashboard/components/Payments/components/PaymentsStatsSection/PaymentStats.section.tsx
@@ -8,7 +8,7 @@ import { fetchApi } from 'src/utils/fetchApi';
 import { useLocalizedActiveCoinValueFormatter } from 'src/hooks/useDisplayReward';
 import { useLocalizedDateFormatter } from 'src/utils/date.utils';
 import { Helmet } from 'react-helmet-async';
-import { MinerPaymentsList } from './MinerPayments.list';
+import { MinerPaymentsList } from '../MinerPaymentsList/MinerPayments.list';
 import { useTranslation } from 'react-i18next';
 import {
   useLocalizedCurrencyFormatter,
