+++ src/pages/MinerDashboard/components/Rewards/components/MinerReportBlocksSection/types.tsx
@@ -1,0 +1,9 @@
+export type ApiBlock = {
+  share: number;
+  reward: number;
+  confirmed: boolean;
+  blockNumber: number;
+  timestamp: number;
+  hash: string;
+  blockType: 'block' | 'uncle' | 'orphan';
+};
