+++ src/pages/MinerDashboard/components/Rewards/components/MinerRewardsStatsSection/MinerRewardStats.section.tsx
